moves withStyles to containers removes dive where applied and modifies tests closes #1948
