+++ __tests__/src/components/WindowTopMenuButton.test.js
@@ -14,7 +14,7 @@ function createWrapper(props) {
       t={str => str}
       {...props}
     />,
-  ).dive(); // unwrap HOC created by withStyles()
+  );
 }
 
 describe('WindowTopMenuButton', () => {
