+++ src/components/WorkspaceControlPanel.js
@@ -1,7 +1,6 @@
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
 import classNames from 'classnames';
-import { withStyles } from '@material-ui/core/styles';
 import Drawer from '@material-ui/core/Drawer';
 import WorkspaceControlPanelButtons
   from '../containers/WorkspaceControlPanelButtons';
@@ -36,17 +35,4 @@ WorkspaceControlPanel.propTypes = {
   classes: PropTypes.object.isRequired, // eslint-disable-line react/forbid-prop-types
 };
 
-/**
- * @private
- */
-const styles = theme => ({
-  ctrlBtn: {
-    margin: theme.spacing.unit,
-  },
-  drawer: {
-    overflowX: 'hidden',
-    height: '100%',
-  },
-});
-
-export default withStyles(styles)(WorkspaceControlPanel);
+export default WorkspaceControlPanel;
