+++ src/containers/ManifestListItemError.js
@@ -1,6 +1,7 @@
 import { compose } from 'redux';
 import { connect } from 'react-redux';
 import { withNamespaces } from 'react-i18next';
+import { withStyles } from '@material-ui/core/styles';
 import { fetchManifest, removeManifest } from '../state/actions/manifest';
 import ManifestListItemError from '../components/ManifestListItemError';
 
@@ -10,10 +11,27 @@ const mapDispatchToProps = {
   onTryAgainClick: fetchManifest,
 };
 
+/**
+ *
+ * @param theme
+ * @returns {{manifestIdText: {wordBreak: string},
+ * errorIcon: {color: string, width: string, height: string}}}
+ */
+const styles = theme => ({
+  errorIcon: {
+    color: theme.palette.error.main,
+    height: '2rem',
+    width: '2rem',
+  },
+  manifestIdText: {
+    wordBreak: 'break-all',
+  },
+});
+
 const enhance = compose(
-  connect(null, mapDispatchToProps),
   withNamespaces(),
-  // further HOC
+  withStyles(styles),
+  connect(null, mapDispatchToProps),
 );
 
 export default enhance(ManifestListItemError);
