+++ src/containers/WindowIcon.js
@@ -1,4 +1,6 @@
 import { connect } from 'react-redux';
+import { withStyles } from '@material-ui/core';
+import { compose } from 'redux';
 import { getWindowManifest, getManifestLogo } from '../state/selectors';
 import WindowIcon from '../components/WindowIcon';
 
@@ -7,4 +9,16 @@ const mapStateToProps = (state, { windowId }) => ({
   manifestLogo: getManifestLogo(getWindowManifest(state, windowId)),
 });
 
-export default connect(mapStateToProps)(WindowIcon);
+const styles = {
+  logo: {
+    height: '2.5rem',
+    paddingRight: 8,
+  },
+};
+
+const enhance = compose(
+  withStyles(styles),
+  connect(mapStateToProps),
+);
+
+export default enhance(WindowIcon);
