+++ src/containers/WindowViewSettings.js
@@ -25,10 +25,9 @@ const mapStateToProps = (state, props) => (
 );
 
 const enhance = compose(
-  connect(mapStateToProps, mapDispatchToProps),
   withNamespaces(),
+  connect(mapStateToProps, mapDispatchToProps),
   miradorWithPlugins,
-  // further HOC go here
 );
 
 export default enhance(WindowViewSettings);
