+++ src/containers/WorkspaceAddButton.js
@@ -1,6 +1,7 @@
 import { connect } from 'react-redux';
 import { compose } from 'redux';
 import { withNamespaces } from 'react-i18next';
+import { withStyles } from '@material-ui/core';
 import * as actions from '../state/actions';
 import miradorWithPlugins from '../lib/miradorWithPlugins';
 import WorkspaceAddButton from '../components/WorkspaceAddButton';
@@ -24,9 +25,21 @@ const mapStateToProps = state => (
  */
 const mapDispatchToProps = { setWorkspaceAddVisibility: actions.setWorkspaceAddVisibility };
 
+/**
+ *
+ * @param theme
+ * @returns {{ctrlBtn: {margin: (number|string)}}}
+ */
+const styles = theme => ({
+  ctrlBtn: {
+    margin: theme.spacing.unit,
+  },
+});
+
 const enhance = compose(
-  connect(mapStateToProps, mapDispatchToProps),
   withNamespaces(),
+  withStyles(styles),
+  connect(mapStateToProps, mapDispatchToProps),
   miradorWithPlugins,
 );
