+++ src/containers/ZoomControls.js
@@ -1,6 +1,7 @@
 import { compose } from 'redux';
 import { connect } from 'react-redux';
 import { withNamespaces } from 'react-i18next';
+import { withStyles } from '@material-ui/core';
 import * as actions from '../state/actions';
 import ZoomControls from '../components/ZoomControls';
 
@@ -16,7 +17,6 @@ const mapStateToProps = (state, props) => (
   }
 );
 
-
 /**
  * mapDispatchToProps - used to hook up connect to action creators
  * @memberof Workspace
@@ -24,10 +24,27 @@ const mapStateToProps = (state, props) => (
  */
 const mapDispatchToProps = { updateViewport: actions.updateViewport };
 
+/**
+ *
+ * @param theme
+ * @returns {{zoom_controls: {position: string, right: number},
+ * ListItem: {paddingBottom: number, paddingTop: number}}}
+ */
+const styles = theme => ({
+  zoom_controls: {
+    position: 'absolute',
+    right: 0,
+  },
+  ListItem: {
+    paddingTop: 0,
+    paddingBottom: 0,
+  },
+});
+
 const enhance = compose(
-  connect(mapStateToProps, mapDispatchToProps),
   withNamespaces(),
-  // further HOC go here
+  withStyles(styles),
+  connect(mapStateToProps, mapDispatchToProps),
 );
 
 export default enhance(ZoomControls);
